20250115-002834

+++ api-v1/todos/src/1_data_sqlite.ts
@@ -6,9 +6,9 @@ export class TodosDataSqlite implements ITodoData {
     constructor() {
         console.log("Using SQLite data provider");
     }
-    async getAll(): Promise<Todos> {
+    async getAll(userid: string): Promise<Todos> {
         return new Promise((resolve, reject) => {
-            this.db.all<Todo>("SELECT * FROM todos", (err, rows) => {
+            this.db.all<Todo>("SELECT * FROM todos WHERE userid = ?", [userid], (err, rows) => {
                 if (err) {
                     reject(err);
                 } else {
@@ -17,9 +17,9 @@ export class TodosDataSqlite implements ITodoData {
             });
         });
     }
-    async getOneById(id: string): Promise<Todo | {}> {
+    async getOneById(userid: string, id: string): Promise<Todo | {}> {
         return new Promise((resolve, reject) => {
-            this.db.get<Todo>("SELECT * FROM todos WHERE id = ?", [id], (err, row) => {
+            this.db.get<Todo>("SELECT * FROM todos WHERE id = ? AND userid = ?", [id, userid], (err, row) => {
                 if (err) {
                     reject(err);
                 }
@@ -27,29 +27,37 @@ export class TodosDataSqlite implements ITodoData {
             });
         });
     }
-    async create(todo: Todo): Promise<Todo | {}> {
+    async create(userid: string, todo: Todo): Promise<Todo | {}> {
         return new Promise((resolve, reject) => {
-            this.db.run("INSERT INTO todos (id, title, description, done) VALUES (?, ?, ?, ?)", [todo.id, todo.title, todo.description, todo.done], (err) => {
-                if (err) {
-                    reject(err);
+            this.db.run(
+                "INSERT INTO todos (id, title, description, done, userid) VALUES (?, ?, ?, ?, ?)",
+                [todo.id, todo.title, todo.description, todo.done, userid],
+                (err) => {
+                    if (err) {
+                        reject(err);
+                    }
+                    resolve(todo);
                 }
-                resolve(todo);
-            });
+            );
         });
     }
-    async update(id: string, todo: Todo): Promise<Todo | {}> {
+    async update(userid: string, id: string, todo: Todo): Promise<Todo | {}> {
         return new Promise((resolve, reject) => {
-            this.db.run("UPDATE todos SET title = ?, description = ?, done = ? WHERE id = ?", [todo.title, todo.description, todo.done, id], (err) => {
-                if (err) {
-                    reject(err);
+            this.db.run(
+                "UPDATE todos SET title = ?, description = ?, done = ? WHERE id = ? AND userid = ?",
+                [todo.title, todo.description, todo.done, id, userid],
+                (err) => {
+                    if (err) {
+                        reject(err);
+                    }
+                    resolve(todo);
                 }
-                resolve(todo);
-            });
+            );
         });
     }
-    async delete(id: string): Promise<Todo | {}> {
+    async delete(userid: string, id: string): Promise<Todo | {}> {
         return new Promise((resolve, reject) => {
-            this.db.run("DELETE FROM todos WHERE id = ?", [id], (err) => {
+            this.db.run("DELETE FROM todos WHERE id = ? AND userid = ?", [id, userid], (err) => {
                 if (err) {
                     reject(err);
                 }
